+++ src/Sentry.Maui/Internal/SentryMauiScreenshotProcessor.cs
@@ -18,6 +18,9 @@ internal class SentryMauiScreenshotProcessor : ISentryEventProcessorWithHint
 
     public SentryEvent? Process(SentryEvent @event, SentryHint hint)
     {
+        // Call back before taking the screenshot if the callback is mentioned and if a Screenshot is attached
+        _options.BeforeCaptureScreenshot();
+
         hint.Attachments.Add(new ScreenshotAttachment(_options));
         return @event;
     }